requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub
langchain-huggingface
# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn
pytorch
streamlit
torch